#68 searchPage컴포넌트서 SearchBar 컴포넌트로 isMobile state 위치 변경
